Creating and printing the Invoice is 90% done! :)
I'm almost done with creating the Invoice, calculating its amount, and printing it to the console. The program already creates and prints an invoice in a neat format! All we need to do is add the delivery charge, the dates, and a few other things and we will be done with creating the invoice! :+1: