CSV Manipulation Problems
Dillon and I tried to manipulate CSV's, writing our own methods to do so. However, we found a very good library called Apache Commons CSV that has those methods already that I hope we could use with ease. It's a lesson that I learned in UCI: don't waste time and effort reinventing the wheel when there is already a wheel out there for us to use.